+++ src/main/java/com/bernardomg/association/member/service/DefaultMemberService.java
@@ -8,11 +8,14 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
-import com.bernardomg.association.member.model.ImmutableMember;
 import com.bernardomg.association.member.model.Member;
-import com.bernardomg.association.member.model.request.MemberQueryRequest;
+import com.bernardomg.association.member.model.mapper.MemberMapper;
+import com.bernardomg.association.member.model.request.MemberCreate;
+import com.bernardomg.association.member.model.request.MemberQuery;
+import com.bernardomg.association.member.model.request.MemberUpdate;
 import com.bernardomg.association.member.persistence.model.PersistentMember;
 import com.bernardomg.association.member.persistence.repository.MemberRepository;
+import com.bernardomg.exception.InvalidIdException;
 
 import lombok.AllArgsConstructor;
 
@@ -26,6 +29,8 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 public final class DefaultMemberService implements MemberService {
 
+    private final MemberMapper     mapper;
+
     /**
      * Member repository.
      */
@@ -33,40 +38,41 @@ public final class DefaultMemberService implements MemberService {
 
     @Override
     @PreAuthorize("hasAuthority('MEMBER:CREATE')")
-    public final Member create(final Member member) {
+    public final Member create(final MemberCreate member) {
         final PersistentMember entity;
         final PersistentMember created;
 
         // TODO: Return error messages for duplicate data
         // TODO: Phone and identifier should be unique or empty
 
-        entity = toEntity(member);
-        entity.setId(null);
+        entity = mapper.toEntity(member);
 
         created = repository.save(entity);
 
-        return toDto(created);
+        return mapper.toDto(created);
     }
 
     @Override
     @PreAuthorize("hasAuthority('MEMBER:DELETE')")
-    public final Boolean delete(final Long id) {
-        // TODO: Handle deleting related data
+    public final void delete(final Long id) {
+        if (!repository.existsById(id)) {
+            throw new InvalidIdException(String.format("Failed delete. No member with id %s", id));
+        }
+
+        // TODO: Forbid deleting when there are relationships
 
         repository.deleteById(id);
-
-        return true;
     }
 
     @Override
     @PreAuthorize("hasAuthority('MEMBER:READ')")
-    public final Iterable<Member> getAll(final MemberQueryRequest sample, final Pageable pageable) {
+    public final Iterable<Member> getAll(final MemberQuery sample, final Pageable pageable) {
         final PersistentMember entity;
 
-        entity = toEntity(sample);
+        entity = mapper.toEntity(sample);
 
         return repository.findAll(Example.of(entity), pageable)
-            .map(this::toDto);
+            .map(mapper::toDto);
     }
 
     @Override
@@ -79,7 +85,7 @@ public final class DefaultMemberService implements MemberService {
         found = repository.findById(id);
 
         if (found.isPresent()) {
-            data = toDto(found.get());
+            data = mapper.toDto(found.get());
             result = Optional.of(data);
         } else {
             result = Optional.empty();
@@ -90,47 +96,19 @@ public final class DefaultMemberService implements MemberService {
 
     @Override
     @PreAuthorize("hasAuthority('MEMBER:UPDATE')")
-    public final Member update(final Long id, final Member member) {
+    public final Member update(final Long id, final MemberUpdate member) {
         final PersistentMember entity;
         final PersistentMember updated;
 
-        entity = toEntity(member);
+        if (!repository.existsById(id)) {
+            throw new InvalidIdException(String.format("Failed update. No member with id %s", id));
+        }
+
+        entity = mapper.toEntity(member);
         entity.setId(id);
 
         updated = repository.save(entity);
-        return toDto(updated);
-    }
-
-    private final Member toDto(final PersistentMember entity) {
-        return ImmutableMember.builder()
-            .id(entity.getId())
-            .name(entity.getName())
-            .surname(entity.getSurname())
-            .identifier(entity.getIdentifier())
-            .phone(entity.getPhone())
-            .active(entity.getActive())
-            .build();
-    }
-
-    private final PersistentMember toEntity(final Member data) {
-        return PersistentMember.builder()
-            .id(data.getId())
-            .name(data.getName())
-            .surname(data.getSurname())
-            .identifier(data.getIdentifier())
-            .phone(data.getPhone())
-            .active(data.getActive())
-            .build();
-    }
-
-    private final PersistentMember toEntity(final MemberQueryRequest data) {
-        return PersistentMember.builder()
-            .name(data.getName())
-            .surname(data.getSurname())
-            .identifier(data.getIdentifier())
-            .phone(data.getPhone())
-            .active(data.getActive())
-            .build();
+        return mapper.toDto(updated);
     }
 
 }
